+++ src/Helper/Coordenadas.java
@@ -1,10 +1,10 @@
 package Helper;
 
 public class Coordenadas {
-	private Integer coordenadaX;
-	private Integer coordenadaY;
+	private Double coordenadaX;
+	private Double coordenadaY;
 	
-	public Coordenadas(Integer coordenadaX, Integer coordenadaY) {
+	public Coordenadas(Double coordenadaX, Double coordenadaY) {
 		this.coordenadaX = coordenadaX;
 		this.coordenadaY = coordenadaY;
 	}
@@ -12,28 +12,28 @@ public class Coordenadas {
 	/**
 	 * @return the coordenadaX
 	 */
-	public Integer getCoordenadaX() {
+	public Double getCoordenadaX() {
 		return coordenadaX;
 	}
 
 	/**
 	 * @param coordenadaX the coordenadaX to set
 	 */
-	public void setCoordenadaX(Integer coordenadaX) {
+	public void setCoordenadaX(Double coordenadaX) {
 		this.coordenadaX = coordenadaX;
 	}
 
 	/**
 	 * @return the coordenadaY
 	 */
-	public Integer getCoordenadaY() {
+	public Double getCoordenadaY() {
 		return coordenadaY;
 	}
 
 	/**
 	 * @param coordenadaY the coordenadaY to set
 	 */
-	public void setCoordenadaY(Integer coordenadaY) {
+	public void setCoordenadaY(Double coordenadaY) {
 		this.coordenadaY = coordenadaY;
 	}
